+++ .gitignore
@@ -133,5 +133,5 @@ dmypy.json
 .DS_Store
 
 # Output
-Output/
+output/
