+++ Tests/profile_fitting_factory_test.py
@@ -1,9 +1,9 @@
 import pytest
-from Code.peak_profile_fitting_factory import PeakProfileFitting_Factory as PPF_factory
-from Code.rietveld import Rietveld
-from Code.poly import Poly
+from src.peak_profile_fitting_factory import PeakProfileFitting_Factory as PPF_factory
+from src.rietveld import Rietveld
+from src.poly import Poly
 import numpy as np
-from Code.strategy import Strategy
+from src.strategy import Strategy
 import pandas as pd
 
 # f = open('Malli_80s.allASC.ASC', 'r')
@@ -13,7 +13,7 @@ import pandas as pd
 # I = I/max(I)
 # spectrum = pd.DataFrame({'x':x,'y':I})
 
-spectrum = pd.read_csv('Code/Input/Malli_80s.csv', skiprows=2, header=None, names=['x','y'])
+spectrum = pd.read_csv('src/input/Malli_80s.csv', skiprows=2, header=None, names=['x','y'])
 spectrum['y'] = spectrum['y']/max(spectrum['y'])
 
 def test_poly_factory():
