+++ src/experimental_technique.py
@@ -14,6 +14,6 @@ class ExperimentalTechnique():
         :return: a dataFrame containing x and y values
         """
         d = os.path.dirname(os.path.realpath(__file__))
-        input_path = os.path.join(d, 'Input', str(filename)) # get the full path of the file
+        input_path = os.path.join(d, 'input', str(filename)) # get the full path of the file
         spectrum = pd.read_csv(input_path, skiprows=2, header=None, names=['x','y']) # read file and store in dataFrame
         return spectrum
